requirements.txt
aiohttp==3.9.1
pyyaml==6.0.2
python-dotenv==1.0.1
git+https://github.com/elliottech/lighter-python.git

